feat1: introduced var in the main class feat2: separated truepenny & equivocator classes
