+++ Person/IPerson.cs
@@ -5,7 +5,7 @@ namespace Person
     interface IPerson
     {
         //property
-        string Name { get; }
+        string Name { get;}
 
         //property
         int Age { get; }
